# lib/Pet.js
class Pet {
  constructor () {
    /**
     * @member {number}
     */
    this.id = undefined
    /**
     * @member {string}
     */
    this.name = undefined
    /**
     * @member {string}
     */
    this.tag = undefined
  }
}

module.exports = Pet