# swaggerui-dist/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Petstore
  license:
    name: MIT
servers:
  - url: /
paths:
  /pets:
    get:
      summary: List all pets
      tags:
        - pets
      responses:
        '200':
          description: A paged array of pets
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/Pets"
        default:
          $ref: '#/components/responses/error'
    post:
      summary: Create a pet
      tags:
        - pets
      requestBody:
        $ref: '#/components/requestBodies/pet'
      responses:
        '201':
          description: Null response
        default:
          $ref: '#/components/responses/error'
  /pets/{petId}:
    get:
      summary: Info for a specific pet
      tags:
        - pets
      parameters:
        - name: petId
          in: path
          required: true
          description: The id of the pet to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
        default:
          $ref: '#/components/responses/error'
    put:
      summary: Update a pet
      tags:
        - pets
      parameters:
        - name: petId
          in: path
          required: true
          description: The id of the pet to retrieve
          schema:
            type: string
      requestBody:
        $ref: '#/components/requestBodies/pet'
      responses:
        '200':
          description: Null response
        default:
          $ref: '#/components/responses/error'
    delete:
      summary: delete a pet
      tags:
        - pets
      parameters:
        - name: petId
          in: path
          required: true
          description: The id of the pet to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Null response
        default:
          $ref: '#/components/responses/error'
components:
  schemas:
    Pet:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        tag:
          type: string
    Pets:
      type: array
      items:
        $ref: "#/components/schemas/Pet"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  requestBodies:
    pet:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Pet'
  responses:
    error:
      description: unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"